test: put aws defaults tests in the same file
